refactor(analysis): split content byte helpers
